Gitignore __pycache__/
When using this plugin, a directory __pycache__/ is created in this git
root. As this plugin is checked out as a submodule in a main dotfiles
repo (user repo), git-status will show that this submodule has modified
content which is not true (this is just cached files).

Thus this repo should ignore __pycache__/